kill searchView
